server/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
python-dotenv
openai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pymilvus
sentence-transformers
